# docs/05-api/openapi.yaml
openapi: 3.1.0
info:
  title: TindaFlow POS API
  version: 1.0.0-draft
  description: 'TindaFlow POS HTTP API (Stage 4 contract). Implements the frozen Stage 2 domain (tag stage-2-baseline, commit 2bc8885) and Stage 3 architecture (tag stage-3-baseline, commit c8a8bbe) without redesigning either. Same-origin session-authenticated SPA API for a self-hosted, LAN-first deployment. This is a design-time contract; no implementation exists yet.'
  contact: {name: ABBADev IT Solutions}
servers:
- {url: /api/v1, description: Same-origin relative base path (store LAN server or VPS)}
...
security:
- cookieAuth: []
paths:
...
  /fiscal-installations/{fiscalInstallationId}:
    get:
      operationId: fiscalInstallationGet
      tags: [FiscalInstallation]
      x-capability: FISCAL_CONFIGURATION_MANAGE
      summary: Get a fiscal installation
      parameters:
      - name: fiscalInstallationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: Fiscal installation
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FiscalInstallation'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    cookieAuth: {type: apiKey, in: cookie, name: tindaflow_session, description: 'Laravel-compatible session cookie (HttpOnly, Secure, SameSite=Strict). All state-changing requests additionally require a valid CSRF token via the standard Laravel double-submit-cookie mechanism (X-XSRF-TOKEN header sourced from the XSRF-TOKEN cookie); this is enforced by global middleware and is not repeated as a parameter on every operation in this document — see api-design.md, Authentication.'}
    terminalCookieAuth: {type: apiKey, in: cookie, name: tindaflow_terminal, description: 'Opaque, high-entropy terminal credential issued at enrollment (ADR-011) and stored as a second HttpOnly/SameSite=Strict cookie (Secure mandatory in production). Looked up by deterministic hash against terminals.credential_hash on every request; a revoked (revoked_at IS NOT NULL) or unmatched credential fails with TERMINAL_REVOKED/TERMINAL_NOT_ENROLLED. Required in addition to, never instead of, cookieAuth on every POS_TERMINAL operation (module-a-auth-terminal-initialization.md §14 Ruling 3) — combined in a single Security Requirement Object so both are conjunctively required.'}
...
  responses:
...
    Unauthorized:
      description: AUTHENTICATION_REQUIRED
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: AUTHORIZATION_DENIED or TERMINAL_NOT_ENROLLED/TERMINAL_REVOKED
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string, description: Stable machine-readable code — see error-catalog.md}
            message: {type: string}
            details: {type: object, additionalProperties: true}
            request_id: {type: string}
...
    Accreditation:
      type: object
      description: 'Supplier/software-level accreditation status — independently effective-dated from PermitToUse per ADR-009 (RMC 72-2025: an expiring Certificate of Accreditation does not itself expire the taxpayer''s PTU). All fields nullable/empty until a real accreditation exists — entering values here does not itself constitute BIR accreditation.

        '
      properties:
        number:
          anyOf:
          - {type: string}
          - {type: 'null'}
        date:
          anyOf:
          - {type: string, format: date}
          - {type: 'null'}
        effective_from:
          anyOf:
          - {type: string, format: date}
          - {type: 'null'}
        effective_to:
          anyOf:
          - {type: string, format: date, description: Null = currently in effect}
          - {type: 'null'}
    PermitToUse:
      type: object
      description: 'Taxpayer/installation-level Permit to Use status — independently effective-dated from Accreditation per ADR-009. Carries the Machine Identification Number (MIN), which is a PTU-lifecycle concept per domain-model.md §2.1, not an accreditation-lifecycle one.'
      properties:
        number:
          anyOf:
          - {type: string}
          - {type: 'null'}
        min:
          anyOf:
          - {type: string, description: Machine Identification Number}
          - {type: 'null'}
        date:
          anyOf:
          - {type: string, format: date}
          - {type: 'null'}
        effective_from:
          anyOf:
          - {type: string, format: date}
          - {type: 'null'}
        effective_to:
          anyOf:
          - {type: string, format: date, description: Null = currently in effect}
          - {type: 'null'}
    FiscalInstallationInput:
      type: object
      required: [deployment_model, software_version]
      properties:
        deployment_model:
          type: string
          enum: [STANDALONE, SERVER_CONNECTED]
        machine_serial_number:
          anyOf:
          - {type: string}
          - {type: 'null'}
        software_version: {type: string}
        accreditation:
          anyOf:
          - allOf:
            - {$ref: '#/components/schemas/Accreditation'}
          - {type: 'null'}
        permit_to_use:
          anyOf:
          - allOf:
            - {$ref: '#/components/schemas/PermitToUse'}
          - {type: 'null'}
    FiscalInstallation:
      description: 'Accreditation and PermitToUse are represented as distinct nested objects, never collapsed into one generic "compliance_status" field — they are independent lifecycles (ADR-009/BIR-012) even though V1''s underlying database representation (two column-pairs vs. two child tables) remains a Stage 5 decision that this HTTP shape does not need to anticipate further.

        '
      allOf:
      - {$ref: '#/components/schemas/FiscalInstallationInput'}
      - type: object
        properties:
          id: {type: string, format: uuid}
          store_id: {type: string, format: uuid}
          installed_at: {type: string, format: date-time}
          superseded_at:
            anyOf:
            - {type: string, format: date-time}
            - {type: 'null'}
          terminals:
            type: array
            items: {type: string, format: uuid}
            description: Terminals currently associated via terminal_fiscal_installation (supports both STANDALONE and SERVER_CONNECTED deployment models, ADR-009/BIR-013).
...